+++ changelog.md
@@ -1,7 +1,18 @@
-20170607
+## v0.2.0
+
+2018-03-18
+
+- fix some bugs when selecting UV2 which doesn't exist
+- add diffuse/lightmap to texface functionnality (Blender Render)
+
+## v0.0.2
+
+2017-06-07
 
 - show total stats
 
-20170606
+## v0.0.1
+
+2017-06-06
 
 - adding mesh selection stats (verts & tri)